Feat: Move claims to its own folder + Licenses + crate structure
