+++ src/environmentd/src/bin/environmentd/main.rs
@@ -612,7 +612,6 @@ fn run(mut args: Args) -> Result<(), anyhow::Error> {
                 decoding_key: DecodingKey::from_rsa_pem(jwk.as_bytes())?,
                 tenant_id,
                 now: mz_ore::now::SYSTEM_TIME.clone(),
-                refresh_before_secs: 60,
                 admin_role,
             }))
         }
